refactor(agent): split keystore into crypt, keyfile, and tokenstore
The keystore package bundled three concerns: low-level AES-GCM/atomic-write
primitives, the data-key file lifecycle, and the token cache. Its name read
as 'a store of keys' and hid where tokens lived. Split it into three flat
packages under agent:

- crypt: generic AES-256-GCM Seal/Open and AtomicWrite, no domain knowledge.
- keyfile: data-key load/create/wrap (KEK derivation via Argon2id) + KeyPath.
- tokenstore: the encrypted token cache (Store) + TokenDir.

keyfile and tokenstore both build on crypt and are otherwise independent
(the data key is the only contract between them). Package names avoid the
generic 'key'/'token', which collide with local variables (token in
server.go handleGet is a real compile clash).
