+++ pkg/controller/agent/crypt/atomicwrite.go
@@ -1,4 +1,4 @@
-package keystore
+package crypt
 
 import (
 	"fmt"
@@ -6,17 +6,17 @@ import (
 	"path/filepath"
 )
 
-// tokenFilePerm is the permission of files written by atomicWrite (current user only).
-const tokenFilePerm os.FileMode = 0o600
+// filePerm is the permission of files written by AtomicWrite (current user only).
+const filePerm os.FileMode = 0o600
 
-// dirPerm is the permission of directories created by atomicWrite (current user only).
+// dirPerm is the permission of directories created by AtomicWrite (current user only).
 const dirPerm os.FileMode = 0o700
 
-// atomicWrite writes data to path atomically: it writes to a temporary file in the
+// AtomicWrite writes data to path atomically: it writes to a temporary file in the
 // same directory, sets its permission to 0600, and renames it into place. A
 // concurrent writer can at worst lose its write but never observe a partial file.
 // The parent directory is created with 0700 if it does not exist.
-func atomicWrite(path string, data []byte) error {
+func AtomicWrite(path string, data []byte) error {
 	dir := filepath.Dir(path)
 	if err := os.MkdirAll(dir, dirPerm); err != nil {
 		return fmt.Errorf("create the directory: %w", err)
@@ -26,7 +26,7 @@ func atomicWrite(path string, data []byte) error {
 		return fmt.Errorf("create a temporary file: %w", err)
 	}
 	tmpName := tmp.Name()
-	if err := tmp.Chmod(tokenFilePerm); err != nil {
+	if err := tmp.Chmod(filePerm); err != nil {
 		_ = tmp.Close()
 		_ = os.Remove(tmpName)
 		return fmt.Errorf("set the temporary file permission: %w", err)
