+++ pkg/controller/agent/keyfile/keyfile_internal_test.go
@@ -1,4 +1,4 @@
-package keystore
+package keyfile
 
 import (
 	"bytes"
