+++ pkg/controller/agent/start.go
@@ -8,7 +8,8 @@ import (
 	"runtime"
 
 	agentapi "github.com/suzuki-shunsuke/ghtkn-go-sdk/ghtkn/backend/agent"
-	"github.com/suzuki-shunsuke/ghtkn/pkg/controller/agent/keystore"
+	"github.com/suzuki-shunsuke/ghtkn/pkg/controller/agent/keyfile"
+	"github.com/suzuki-shunsuke/ghtkn/pkg/controller/agent/tokenstore"
 )
 
 // Start runs the agent server in the foreground.
@@ -21,11 +22,11 @@ import (
 // handling is set up by urfave.Main (see cmd/ghtkn/main.go), so this function
 // does not register its own signal handler.
 func (c *Controller) Start(ctx context.Context, logger *slog.Logger) error {
-	keyFile, err := keystore.KeyPath(os.Getenv, runtime.GOOS)
+	keyFile, err := keyfile.KeyPath(os.Getenv, runtime.GOOS)
 	if err != nil {
 		return err //nolint:wrapcheck
 	}
-	dir, err := keystore.TokenDir(os.Getenv, runtime.GOOS)
+	dir, err := tokenstore.TokenDir(os.Getenv, runtime.GOOS)
 	if err != nil {
 		return err //nolint:wrapcheck
 	}
