+++ pkg/controller/agent/tokenstore/paths_internal_test.go
@@ -1,4 +1,4 @@
-package keystore
+package tokenstore
 
 import (
 	"path/filepath"
